Refactor some algorithms to improve efficiency and/or cleanup
## Major Changes
* Convert the loaded_plugin list to use insertion sort plus binary search. This is a straight speed for space tradeoff.
* Cleanup some dead code.
* Replace some dicey macros in ncjson.
* Refactor libdispatch/dutil.c functions to have their own header: ncutil.h.
* Refactor the byte swapping code.
* Move to consistently type HDF5 filter ids as int to correspond with H5Z_filter_t. This change excludes ids in the netcdf.h API.
* Make Zarr format V3 actually use the "bytes" endian handling codec.
* Fix a weird problem with NC_mktmp.

+++ v3_nczarr_test/test_chunking.c
@@ -119,14 +119,14 @@ main(int argc, char **argv)
 	size_t start[2] = {0,0};
 	size_t count[2];
 	for(i=0;i<nvals;i++) {
-	    ivar_data[i] = i;	   
+	    ivar_data[i] = (int)i;
 	}
 	count[0] = UNLIM_SIZE;	
 	count[1] = dim_lens[0];
 	if (nc_put_vara(ncid, ivarid, start, count, ivar_data)) LERR;
     } else {
 	for(i=0; i < NVALS; i++) {
-	    ivar_data[i] = i;
+	    ivar_data[i] = (int)i;
 	}
 	if (nc_put_var(ncid, ivarid, ivar_data)) LERR;
     }
