+++ v3_nczarr_test/test_filter_vlen.c
@@ -94,7 +94,7 @@ defvar(nc_type xtype)
         snprintf(dimname,sizeof(dimname),"dim%u",(unsigned)i);
         CHECK(nc_def_dim(ncid, dimname, dimsize[i], &dimids[i]));
     }
-    CHECK(nc_def_var(ncid, "var", xtype, ndims, dimids, &varid));
+    CHECK(nc_def_var(ncid, "var", xtype, (int)ndims, dimids, &varid));
     return NC_NOERR;
 }
